Removed the Terraform conventions by inlining them into the command.
