+++ source/Calamari.Terraform/DestroyPlanCommand.cs
@@ -5,11 +5,13 @@ using Calamari.Integration.Processes;
 namespace Calamari.Terraform
 {
     [Command("destroyplan-terraform", Description = "Plans the destruction of Terraform resources")]
-    public class DestroyPlanCommand : TerraformCommand
+    public class DestroyPlanCommand : PlanCommand
     {
         public DestroyPlanCommand(ILog log, IVariables variables, ICalamariFileSystem fileSystem, ICommandLineRunner commandLineRunner) 
-            : base(log, variables, fileSystem, new DestroyPlanTerraformConvention(log, fileSystem, commandLineRunner))
+            : base(log, variables, fileSystem, commandLineRunner)
         {
         }
+
+        protected override string ExtraParameter => "-destroy";
     }
 }
